fix: allow metadata to write more than BLOCK_SIZE and fix according journal logic

+++ datanode/buffmgr/bufferManager.go
@@ -1,27 +1,11 @@
 package buffmgr
 
-//var buffermanager *BufferManager
-//
-//type BufferManager struct {
-//	blankBlock []byte
-//	blankFile  []byte
-//}
-
-func GetPaddedBytes(s string, size int) []byte {
+func GetPaddedBytes(s string, size uint64) []byte {
 	// Fill padding with padByte.
 	padByte := []byte(s)[0]
 	paddedData := make([]byte, size)
-	for i := 0; i < size; i++ {
+	for i := uint64(0); i < size; i++ {
 		paddedData[i] = padByte
 	}
 	return paddedData
 }
-
-//func init() {
-//	if buffermanager == nil {
-//		buffermanager = &BufferManager{
-//			GetPaddedBytes(" ", config.BLOCK_SIZE),
-//			GetPaddedBytes(" ", config.FILE_SIZE),
-//		}
-//	}
-//}
